Fix colour issues in combination with SDL/OpenGL
OpenGL expects BGRA but on Mac OS X with a Intel CPU, it seems that  Big
Endianess is assumed by SDL, probably for backward compatibility to the
PowerPC CPUs.

On Linux however, this doesn't occur. So GL_BGRA is appropriate here
and works well.

To clarify this, the calculated values per OS:

- Linux:

  What we want: BGRA
  What we get in memory: ARGB
  Shifts: RGB (19, 11, 3)

- Mac OS X:

  What we want: BGRA
  What we get in memory: BGRA
  Shifts: BGR (19, 11, 3)

This problem is fixed by using SDL_OPENGLBLIT, which does an alignment
check in SDL and results in correctly positioned shift values for each
endianess. At the end, the Alpha channel is shifted in a way so that
it fits the memory format expected by OpenGL: BGRA.